Added routing for profile
Added routing for the profile element. Since I do not want to intrude on someone else's work for now, I did not complete the profile handlebars. This will need to be changed when signup and login is fully implemented, but for now, this will do as a proof of concept.